create module for cache operations / refactor / not tested
