+++ burn_after/burn_after/core/redis_helper.py
@@ -11,23 +11,38 @@ cache = redis.StrictRedis(
     decode_responses=True  # чтобы строки возвращались как str, а не bytes
 )
 
-# Функция для сериализации и записи данных в кеш
-def cache_set_json(key, value: dict, ex=None):
+
+def set_json(key, value: dict, ex=None):
     """Сериализует значение и сохраняет в кеш."""
     cache.set(key, json.dumps(value), ex=ex)
 
-# Функция для чтения данных из кеша и десериализации
-def cache_get_json(key):
+def get_json(key):
     """Читает значение из кеша и десериализует его."""
     data = cache.get(key)
     return json.loads(data) if data else None
 
-# Функция для удаления данных из кеша
-def cache_delete(key):
+
+def delete(key):
     """Удаляет ключ из кеша."""
     cache.delete(key)
 
-# Функция для получения длины Zset
+
 def get_zset_length(zset_key):
     """Возвращает длину Zset в Redis."""
     return cache.zcard(zset_key)
+
+def get_zset_segment(zset_key, start: int, end: int, sort_reflection: bool) -> list:
+    """Возвращает отрезок элемнтов по старту и концу из zset"""
+    if sort_reflection:
+        return cache.zrevrange(zset_key, start, end)
+    else:
+        return cache.zrange(zset_key, start, end)
+        
+def exists(key) -> bool:
+    """проверяет наличие ключа в кеше"""
+    return cache.exists(key)
+
+def set_zset(zset_key, mapping: dict):
+    """создает zset по ключу с нужной разметкой"""
+    cache.zadd(zset_key, mapping=mapping)
+
