+++ burn_after/posts/serializers.py
@@ -1,6 +1,5 @@
 from rest_framework import serializers
 from .models import Post
-# from .redis import get_categories_from_cache
 
 class PostQueryParamsSerializer(serializers.Serializer):
     category = serializers.CharField(required=True, allow_blank=False)
@@ -17,7 +16,9 @@ class PostQueryParamsSerializer(serializers.Serializer):
         return value
 
     def validate_category(self, value):
-        from .redis import get_categories_from_cache
+        
+        from .cache.categories import get_categories_from_cache
+        
         category = value[0] if isinstance(value, list) else value
         if not category in get_categories_from_cache():
             raise serializers.ValidationError(f"Category with name{value} doesn't exist")
